feat(iteration): add Repeat function to repeat a given character a specified number of times
This commit adds a new file `repeat.go` in the `iteration` package. The file contains a `Repeat` function that takes a character and a count as input and returns a string with the character repeated the specified number of times.

A new file `repeat_test.go` is also added in the same package. It contains a test function `TestRepeat` that tests the `Repeat` function by checking if the repeated string matches the expected value. Additionally, a benchmark function `BenchmarkRepeat` is included to measure the performance of the `Repeat` function.

These changes are made to provide a reusable function for repeating characters in the `iteration` package.